requirements.txt

pydantic>=2.5
sounddevice>=0.4
azure-cognitiveservices-speech>=1.37.0
aiohttp>=3.9
python-dotenv>=1.0
rich>=13.7
sqlalchemy>=2.0
openai>=1.0
numpy>=1.26
